test: rename tests to preserve order
